requirements.txt
certifi==2022.9.24
chardet==5.0.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Caching==2.0.1
idna==3.4.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
urllib3==1.26.13
MarkupSafe==2.1.1
Werkzeug==2.2.2
redis==3.5.3
dnspython==2.2.1